Perform a more thorough check on ASCII STL files.
Some binary STL files can include "solid" in the first 80 characters. This causes the current code to segfault, when it tries to treat a binary STL as an ASCII STL. This change takes advantage of the info that the smallest possible STL file is 186 bytes, and reads in a "tail" of that length from the end of the file, and searches it for "endsolid". If found, then it is legitimate ASCII. If not, then it's either binary or broken/weird.