Correctly handle object name in bytearray data
The STL data may come as text or bytearray depending on the source.

The STL output is using string formatting to output numbers, and later
converts the formatted numbers to bytearray to write out.

This does not work for the object name when it comes as byte array. Wehn
formatted as string the extra b'' is added around it.

Change the p() functio so that it can accept a byte array, and pass in
the name as byte array. Convert the name to byte array before passing it
to p() in the case it's text.

This avoids the superfluous conversion.

Fixes: #233